+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-09-13
+### Docs
+- Improved README a lot
+- Improved in-code documentation
+
+### Feat
+- Add support for Quest start / complete conditions through the use of IQuestCondition
+- Add OnStepStartConditionsNotMet & OnStepCompleteConditionsNotMet events to the QuestManager & as relayed UnityEvents in the QuestManagerEventHub.
+- Add optional support for a reward system with IQuestReward lists applicable to Quest Steps, Quests & Quest Lines
+
+### Fix
+- Clarified the event sytem and event orders
+- Changed the ShowMissingRequirements params to be a QuestEventInfo to line up all the events and not have too much differences in what the events provide
+
 ## [1.0.0] - 2024-09-12
 ### Feat
 First release. The logic behind the system is fine for a besis to expand on.